Remove JSONType, change ObjectLike/ArrayLike to objectlike/arraylike.
This makes the interface a bit simpler, I believe. It also removes
the JSONType vs. JSONTypes internal confusion.

Though objectlike/arraylike are now more similar to dictlike,
I realized they aren't really related. On the materialization side,
we have `mutable`, `kwdef`, and `dictlike`; and it's only on the
output side that we consider objectlike/arraylike.