Fix some checkendpos logic
The issue here is that we need to track whether a lazy object is
actually the root object or not to properly do a checkendpos.
We were failing a simple case like " 123a" where the number
didn't technically start at byte position 1.
We were also erroring on lazy arrays in jsonlines format, since
their first element can start at position 1, but the first element
isn't the root array.