Fairly large refactoring of the core applyeach interface.
* Get rid of Continue in favor of the opposite kind of api
* i.e. instead of requiring function arguments to applyeach to return
  Continue in order to avoid short-circuiting, we flip it around and
  say short-circuiting is only achieved if `f` returns a value wrapped
  in an `EarlyReturn`.
* This is an easy enough change, but we still want to be able to
  communicate state from `f` back to the core parsing applyobject/
  applyarray functions, so we introduce an `UpdatedState` internal
  type that we special-case in applyobject/applyarray and that
  our materialize closures return.

The overall effect of this refactoring is that it's much more
convenient to use `JSONBase.applyeach` as a user, since I don't
have to fiddle around with remembering to always return `Continue`.
i.e. it acts a lot more like `Base.foreach` in that return values
are ignored by default. It just has the enhanced functionality that
returning EarlyReturn will trigger short-circuiting.